QueryBuilder was moved up to QueryController (before named QueryContainer)
